+++ middlewares/validateProducts.js
