+++ models/productModel.js
@@ -2,7 +2,7 @@ export const products = [];
 import mongoose from "mongoose";
 import { toJSON } from "@reis/mongoose-to-json";
 
-const advertSchema = new mongoose.Schema(
+const productSchema = new mongoose.Schema(
   {
     title: { type: String, required: true },
 
@@ -19,8 +19,8 @@ const advertSchema = new mongoose.Schema(
   { timestamps: true },
 );
 
-advertSchema.plugin(toJSON);
+productSchema.plugin(toJSON);
 
-const Product = mongoose.model("Product", advertSchema);
+const Product = mongoose.model("Product", productSchema);
 
 export default Product;
